+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.2.0] - 2025-03-17
+
 ### Added
 
 - Make “German language skills for international students” taxonomy hierarchical.
@@ -142,7 +144,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[Unreleased]: https://github.com/RRZE-Webteam/FAU-Studium/compare/2.1.0...HEAD
+[Unreleased]: https://github.com/RRZE-Webteam/FAU-Studium/compare/2.2.0...HEAD
+[2.2.0]: https://github.com/RRZE-Webteam/FAU-Studium/compare/2.1.0...2.2.0
 [2.1.0]: https://github.com/RRZE-Webteam/FAU-Studium/compare/2.0.1...2.1.0
 [2.0.1]: https://github.com/RRZE-Webteam/FAU-Studium/compare/2.0.0...2.0.1
 [2.0.0]: https://github.com/RRZE-Webteam/FAU-Studium/compare/1.2.7...2.0.0